ai-service/requirements.txt
flask==3.0.3
ultralytics==8.3.40
roboflow==1.1.48
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
